+++ src/rads_toolbox/commands/switch_python_version.py
@@ -4,7 +4,8 @@ import click
 
 from rads_toolbox.constants import PackageManager
 from rads_toolbox.contexts import AppContext, pass_app_context
-from rads_toolbox.utils import OnError, print_header, run
+from rads_toolbox.execution import OnError, run
+from rads_toolbox.terminal_output import print_header
 
 
 @click.command()